feaat: first draft refactor
