+++ src/spectre_core/receivers/plugins/_gr.py
